refactor(typedoc): move files
