+++ packages/typedoc/generators.json
@@ -3,10 +3,15 @@
   "name": "typedoc",
   "version": "0.0.1",
   "generators": {
-    "typedoc": {
+    "_typedoc": {
       "factory": "./src/generators/typedoc/generator",
       "schema": "./src/generators/typedoc/schema.json",
-      "description": "typedoc generator"
+      "description": "Generator for Typedoc targets."
+    },
+    "config": {
+      "factory": "./src/generators/typedoc/generator",
+      "schema": "./src/generators/typedoc/schema.json",
+      "description": "Generator for Typedoc targets."
     }
   }
 }
