--- packages/typedoc/src/generators/process/configureProject/getExecutorName.ts
@@ -1,4 +1,0 @@
-import * as pkg from '../../../../package.json'
-import * as project from '../../../../project.json'
-
-export const getExecutorName = (): string => `${pkg.name}:${project.name}`
