+++ packages/typedoc/src/generators/typedoc/files/src/typedoc.js__template__
@@ -1,0 +1,1 @@
+<%- JSON.stringify(options, null, 2) %>
