+++ packages/typedoc/src/generators/typedoc/generator.ts
@@ -1,6 +1,6 @@
 import { Tree } from '@nrwl/devkit'
 import { TypedocGeneratorSchema } from './schema'
-import { addDependencies, registerOperation, configureProject, configureTypedoc } from '../process'
+import { addDependencies, registerOperation, configureProject, configureTypedoc } from './process'
 
 export default async function (tree: Tree, options: TypedocGeneratorSchema) {
   const installDependencies = addDependencies(tree)
