+++ packages/typedoc/src/generators/typedoc/process/configureProject/configureProject.ts
@@ -1,8 +1,8 @@
 import { joinPathFragments, normalizePath, Tree } from '@nrwl/devkit'
-import { TypedocGeneratorSchema } from '../../typedoc/schema'
+import { TypedocGeneratorSchema } from '../../schema'
 import { readProjectConfiguration, updateProjectConfiguration } from '@nrwl/devkit'
 import { getExecutorName } from './getExecutorName'
-import { name } from '../../../../project.json'
+import { name } from '../../../../../project.json'
 
 export const configureProject = (tree: Tree, options: TypedocGeneratorSchema): void => {
   const config = readProjectConfiguration(tree, options.project)
