+++ packages/typedoc/src/generators/typedoc/process/configureProject/getExecutorName.ts
@@ -1,0 +1,4 @@
+import * as pkg from '../../../../../package.json'
+import * as project from '../../../../../project.json'
+
+export const getExecutorName = (): string => `${pkg.name}:${project.name}`
