+++ packages/typedoc/src/generators/typedoc/process/configureProject/getTsConfig.ts
